Code review + File swap
